+++ lib/danger/ci_source/drone.rb
@@ -1,0 +1,18 @@
+# http://readme.drone.io/usage/variables/
+
+module Danger
+  module CISource
+    class Drone < CI
+      def self.validates?(env)
+        return !env["DRONE"].nil?
+      end
+
+      def initialize(env)
+        self.repo_slug = env["DRONE_REPO"]
+        if env["DRONE_PULL_REQUEST"].to_i > 0
+          self.pull_request_id = env["DRONE_PULL_REQUEST"]
+        end
+      end
+    end
+  end
+end